Implement ButtonBase (EnhancedButton... but slightly less enhanced)
