+++ docs/api/Button/Button.md
@@ -1,7 +1,18 @@
 Button
 ======
 
-Short button description goes here
+Buttons communicate the action that will occur when the user
+touches them.
+
+Material buttons trigger an ink reaction on press. They may display
+text, imagery, or both. Flat buttons and raised buttons are the
+most commonly used types.
+
+```js
+import Button from 'material-ui/Button';
+
+const Component = () => <Button>Hello World</Button>;
+```
 
 Props
 -----
